MaatruCare_ML/MoodAnalyser/requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentencepiece>=0.1.99
accelerate>=0.24.0
protobuf
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pymongo>=4.6.0
reportlab>=4.0.0
python-dotenv>=1.0.0
report_generator